+++ src/templatemanager/templates/datePicker/datePicker.tsx
@@ -28,22 +28,23 @@ export function DatePickerExt(props: any) {
     const dateFormats: any = {
         'DD-MM-YYYY': 'dd-MM-yyyy',
         'MM-DD-YYYY': 'MM-dd-yyyy',
-        'YYYY-MM-DD': 'yyyy-MM-dd',
-        'YYYY-DD-MM': 'yyyy-dd-MM'
+        'YYYY-MM-DD': 'yyyy-MM-dd'
     }
 
     const getConvertedDate = (date: any, format: any) => {
-        if (typeof date == 'string') {
-            date = new Date(date);
+        if (date instanceof Date && !isNaN(date.getTime())) {
+            return date;
         }
-        const day = String(date.getDate()).padStart(2, '0');
-        const month = String(date.getMonth() + 1).padStart(2, '0');
-        const year = date.getFullYear();
-        const formattedDate = format
-            .replace('DD', day)
-            .replace('MM', month)
-            .replace('YYYY', year);
-        return formattedDate;
+        const map: any = {
+            'DD-MM-YYYY': [0, 1, 2],
+            'MM-DD-YYYY': [1, 0, 2],
+            'YYYY-MM-DD': [2, 1, 0]
+        };
+    
+        const idx = map[format];
+        const parts = date.split('-');
+        const [d, m, y] = idx.map((i: number) => Number(parts[i]));
+        return new Date(y, m - 1, d);
     }
 
     useEffect(() => {
@@ -96,22 +97,23 @@ export function DatePickerInline(props: any) {
     const dateFormats: any = {
         'DD-MM-YYYY': 'dd-MM-yyyy',
         'MM-DD-YYYY': 'MM-dd-yyyy',
-        'YYYY-MM-DD': 'yyyy-MM-dd',
-        'YYYY-DD-MM': 'yyyy-dd-MM'
+        'YYYY-MM-DD': 'yyyy-MM-dd'
     }
 
     const getConvertedDate = (date: any, format: any) => {
-        if (typeof date == 'string') {
-            date = new Date(date);
+        if (date instanceof Date && !isNaN(date.getTime())) {
+            return date;
         }
-        const day = String(date.getDate()).padStart(2, '0');
-        const month = String(date.getMonth() + 1).padStart(2, '0');
-        const year = date.getFullYear();
-        const formattedDate = format
-            .replace('DD', day)
-            .replace('MM', month)
-            .replace('YYYY', year);
-        return formattedDate;
+        const map: any = {
+            'DD-MM-YYYY': [0, 1, 2],
+            'MM-DD-YYYY': [1, 0, 2],
+            'YYYY-MM-DD': [2, 1, 0]
+        };
+    
+        const idx = map[format];
+        const parts = date.split('-');
+        const [d, m, y] = idx.map((i: number) => Number(parts[i]));
+        return new Date(y, m - 1, d);
     }
 
     const handleSubmit = () => {